Improve TextVersion test
